refactor: factor out IsUninhabited into its own file
